requirements-prod.txt
# ===== CORE DJANGO =====
Django==5.2.8
python-dotenv==1.0.0

# ===== DATABASE =====
# SQLite (included with Python, mais pour PROD: utiliser PostgreSQL)
psycopg2-binary==2.9.9          # PostgreSQL adapter pour Django
# OU pour MySQL:
# mysqlclient==2.2.0

# ===== IMAGE PROCESSING =====
Pillow==10.1.0                   # Pour traitement images (PaymentReceipt.receipt_image)
qrcode[pil]==7.4.2              # Pour génération QR codes

# ===== WEB SERVER =====
gunicorn==21.2.0                 # WSGI HTTP Server (pour production)
whitenoise==6.6.0                # Serveur fichiers statiques efficace

# ===== EMAIL =====
# Django inclut SMTP, mais optionnel pour Celery async:
# celery==5.3.4                   # Task queue asynchrone
# redis==5.0.1                    # Cache + message broker pour Celery

# ===== SECURITY =====
# Django inclut CSRF, XSS protection
cryptography==41.0.7             # Pour chiffrement optionnel PII

# ===== MONITORING & LOGGING =====
# sentry-sdk==1.38.0              # Crash reporting (optionnel)

# ===== UTILITIES =====
requests==2.31.0                 # HTTP client
pytz==2023.3.post1              # Timezone support

# ===== DEVELOPMENT (ne pas installer en prod) =====
# Décommenter uniquement en développement:
# django-extensions==3.2.3
# django-debug-toolbar==4.2.0
